+++ modules/core/src/main/java/org/locationtech/jts/triangulate/polygon/PolygonEarClipper.java
@@ -9,7 +9,7 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package org.locationtech.jts.triangulatepoly;
+package org.locationtech.jts.triangulate.polygon;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,7 +22,7 @@ import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.Polygon;
 import org.locationtech.jts.geom.Triangle;
-import org.locationtech.jts.triangulatepoly.tri.Tri;
+import org.locationtech.jts.triangulate.tri.Tri;
 
 /**
  * Triangulates a polygon using the Ear-Clipping technique.
