+++ modules/core/src/main/java/org/locationtech/jts/triangulate/polygon/VertexSequencePackedRtree.java
@@ -9,7 +9,7 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package org.locationtech.jts.triangulatepoly;
+package org.locationtech.jts.triangulate.polygon;
 
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
