+++ modules/core/src/test/java/org/locationtech/jts/triangulate/polygon/PolygonTriangulatorTest.java
@@ -9,9 +9,10 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package org.locationtech.jts.triangulatepoly;
+package org.locationtech.jts.triangulate.polygon;
 
 import org.locationtech.jts.geom.Geometry;
+import org.locationtech.jts.triangulate.polygon.PolygonTriangulator;
 
 import junit.textui.TestRunner;
 import test.jts.GeometryTestCase;
